+++ test/libsolidity/syntaxTests/inlineAssembly/pc.sol
@@ -1,0 +1,9 @@
+contract C {
+	function f() pure public {
+		assembly {
+			pop(pc())
+		}
+	}
+}
+// ----
+// Warning: (61-63): The "pc" instruction is deprecated and will be removed in the next breaking release.